# Rectangling function rewrite