… NEWS.md
… tests/testthat/_snaps/rectangle.md