`unnest_longer()`:
Closes #1201
Closes #1199 
Closes #1198 
Closes #1196 
Closes #1195 
Closes #1194 
Closes #1193 
Closes #1034 
Closes #1029 
Closes #740 

`unnest_wider()`:
Closes #1188 
Closes #1187 
Closes #1186
Closes #1125 
Closes #1133 
Closes #1060 

`hoist()`:
Added test of current behavior for #1203 
Closes #1044 
Closes #1000 
Closes #999 
Closes #998 
Closes #996 

Other:
Part of #1185 

---

Affected revdeps:

- ffscrapr
  - Pre-emptive PR already merged https://github.com/ffverse/ffscrapr/pull/344

- wehoop
  - Lots of off-label usage of `unnest_wider()`. PR https://github.com/saiemgilani/wehoop/pull/14
  - Two tests explicitly check for a set of column names after all the unnesting is applied. Fixing the first part of https://github.com/tidyverse/tidyr/issues/1125#issuecomment-959802379 means that columns with all empty elements are now (correctly) retained, which means this set of column names should be updated. Luckily this test isn't tested on CRAN, so it should be easy. PR https://github.com/saiemgilani/wehoop/pull/15

Update: Both revdeps have already merged the PRs and have updated their packages on CRAN. So there are now no affected revdeps.

---

This PR completely rewrites the rectangling functions. They are now much more consistent in a variety of edge cases, and are often much faster than they used to be. I anticipate very little breaking changes, except in some extreme off-label usage situations. I will leave in-line comments, but I think this generally should be reviewed as if these were completely new functions.

Here is the most interesting performance difference I found so far, which comes from https://stackoverflow.com/questions/68897169/any-speedier-alternatives-to-tidyrunnest-longer-when-dealing-with-nested-nam. Mainly this comes from getting rid of `tibble()` in favor of `new_data_frame()`. I think #1133 also identified this. (It is also worth noting that in this case a simple `unnest()` is even faster. Like, in the ~200ms range. But it is nice to see this improvement).

```r
library(tidyr)

elt <- 1:5
col <- rep(list(elt), 200000L)

df <- tibble(col = col)
df
#> # A tibble: 200,000 × 1
#>    col      
#>    <list>   
#>  1 <int [5]>
#>  2 <int [5]>
#>  3 <int [5]>
#>  4 <int [5]>
#>  5 <int [5]>
#>  6 <int [5]>
#>  7 <int [5]>
#>  8 <int [5]>
#>  9 <int [5]>
#> 10 <int [5]>
#> # … with 199,990 more rows

# CRAN
bench::mark(unnest_longer(df, col))
#> # A tibble: 1 × 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest_longer(df, col)    1.14m    1.14m    0.0146     292MB     2.09

# This PR
bench::mark(unnest_longer(df, col))
#> # A tibble: 1 × 6
#>   expression                  min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>             <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 unnest_longer(df, col)    2.45s    2.45s     0.408    39.6MB     4.90
```