@mgirlich 2021-11-10
So far I only had a look at `unnest_longer()` and `df_unchop()`. The code looks good to me and easy to read. Great work!

Two things I found:
1. Partial inner names cause an error

``` r
devtools::load_all("~/GitHub/tidyr/")
#> ℹ Loading tidyr
tibble(y = list(a = 1:2, 3:4)) %>% 
  unnest_longer(y, indices_include = FALSE)
#> Error: If any elements of `.data` are named, all must be named
```

2. The names of the `ptype` and `transform` argument should be checked with `vec_as_names(repair = "check_unique")`

``` r
devtools::load_all("~/GitHub/tidyr/")
#> ℹ Loading tidyr
# two ptypes for x
tibble(x = integer()) %>% 
  unnest_longer(
    c(x),
    ptype = list(x = double(), x = integer())
  )
#> # A tibble: 0 × 1
#> # … with 1 variable: x <dbl>

# unnamed list
tibble(x = integer()) %>% 
  unnest_longer(
    c(x),
    ptype = list(double())
  )
#> # A tibble: 0 × 1
#> # … with 1 variable: x <int>

# maybe warn/inform about unused ptypes?
tibble(x = integer()) %>% 
  unnest_longer(
    c(x),
    ptype = list(y = "a")
  )
#> # A tibble: 0 × 1
#> # … with 1 variable: x <int>
```

<sup>Created on 2021-11-10 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>