@mgirlich 2021-11-10
For `unnest_wider()`

1. Add a test with a data frame column and `names_sep`
```r
tibble(
  y = tibble(a = 1, b = 11)
) %>% 
  unnest_wider(y, names_sep = "_")
```

2. Add a test for a column that is not a list and not a data frame

```r
tibble(
  y = c(a = 1, b = 11)
) %>% 
  unnest_wider(y, names_sep = "_")
```