@DavisVaughan 2021-11-10
Thanks @mgirlich, very helpful!

1. This is due to an annoying list-of bug that would be fixed by https://github.com/r-lib/vctrs/pull/784. I'll try and get that in the next vctrs release, it has affected me many times now. I've worked around it the best I can with `tidyr_temporary_new_list_of()`. There is one edge case that it affects, which I opened an issue for #1212.

2. I added more input validation here, but I think unused ptype/transform elements are going to have to go by silently. I don't think it is worth it to check for these.

3. I think your intuition is right. When converting a non-list to a list, there are two things you can do with the names. You can move them to the outer list (this is what we currently do, and what CRAN tidyr does but doesn't have tests for) or you can keep them on the inner elements after converting to a list. The latter aligns with your suggestion, and I now feel that that is correct because it uses more standard vctrs tooling (i.e. `vec_chop()` rather than `tidyr_chop()`) and it generally results in more expected output, even though these are non-primary data types. Below are the two options. I've switched over to the `equivalent2` approach you suggested.

``` r
library(tidyr)
library(vctrs)

# Say we have this:
df <- tibble(x = c(a = 1L, b = 2L))

# Names are moved to outer list (tidyr_chop())
equivalent1 <- tibble(x = list_of(a = 1L, b = 2L))

# Names are kept on inner elements (vec_chop())
equivalent2 <- tibble(x = list_of(c(a = 1L), c(b = 2L)))

unnest_wider(equivalent1, x)
#> New names:
#> * `` -> ...1
#> New names:
#> * `` -> ...1
#> # A tibble: 2 × 1
#>    ...1
#>   <int>
#> 1     1
#> 2     2

unnest_wider(equivalent2, x)
#> # A tibble: 2 × 2
#>       a     b
#>   <int> <int>
#> 1     1    NA
#> 2    NA     2

longer1 <- unnest_longer(equivalent1, x)
longer1
#> # A tibble: 2 × 1
#>       x
#>   <int>
#> 1     1
#> 2     2

longer1$x
#> [1] 1 2

longer2 <- unnest_longer(equivalent2, x)
longer2
#> # A tibble: 2 × 2
#>       x x_id 
#>   <int> <chr>
#> 1     1 a    
#> 2     2 b

longer2$x
#> a b 
#> 1 2
```

4. I've added those two tests, thanks!