@mgirlich 2021-11-11
> I added more input validation here, but I think unused ptype/transform elements are going to have to go by silently. I don't think it is worth it to check for these.

Unused `ptype`/`transform` element might just be a typo. On the other hand in `unnest_wider()` unused `ptype` elements might be used for something different:

Currently, `unnest_wider()` isn't stable regarding the shape of the output:

``` r
library(tidyr)
df <- tibble(
  x = 1:2,
  y = list(
    list(a = 1),
    list(a = 2, b = 2)
  )
)

df %>% unnest_wider(y)
#> # A tibble: 2 × 3
#>       x     a     b
#>   <int> <dbl> <dbl>
#> 1     1     1    NA
#> 2     2     2     2

df %>% dplyr::slice(1) %>% unnest_wider(y)
#> # A tibble: 1 × 2
#>       x     a
#>   <int> <dbl>
#> 1     1     1
```

<sup>Created on 2021-11-11 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>

The `ptype` argument can be used to ensure the types of the columns. Maybe one could also use it to ensure to also have the columns `x`, `a`, and `b` in the second case:

```r
df %>% dplyr::slice(1) %>% unnest_wider(y, ptype = list(a = double(), b = double()))
#> # A tibble: 1 × 2
#>       x     a
#>   <int> <dbl>
#> 1     1     1

# maybe this should output
#> # A tibble: 1 × 3
#>       x     a     b
#>   <int> <dbl> <dbl>
#> 1     1     1    NA
```