@mgirlich 2021-11-11
I am still not entirely sure about the handling of indices as they are not type stable:

``` r
devtools::load_all("~/GitHub/tidyr/")
#> ℹ Loading tidyr
df <- tibble(x = list(c(1, 2), c(a = 3)))

df %>% 
  unnest_longer(x, indices_include = TRUE)
#> # A tibble: 3 × 2
#>       x x_id 
#>   <dbl> <chr>
#> 1     1 <NA> 
#> 2     2 <NA> 
#> 3     3 a

df %>% 
  dplyr::slice(1) %>% 
  unnest_longer(x, indices_include = TRUE)
#> # A tibble: 2 × 2
#>       x  x_id
#>   <dbl> <int>
#> 1     1     1
#> 2     2     2
```

<sup>Created on 2021-11-11 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>

I am not sure whether this is a common scenario but it feels like this goes against the idea of the shape and type predictability in the tidyverse.