@DavisVaughan 2021-11-11
> Maybe one could also use it to ensure to also have the columns x, a, and b in the second case:

I think that would stray pretty far away from how `ptype` is used elsewhere in tidyr. But that could be the place of a "spec" object, like mentioned in https://github.com/tidyverse/tidyr/issues/835. I doubt we add that in this release, but that seems like the right place for this idea

> I am still not entirely sure about the handling of indices as they are not type stable:

Yea I think we are probably stuck with this at this point for backwards compatibility's sake. I agree that some alternative API might have been cleaner / more type stable (like `names_to` and `indices_to` you mentioned in https://github.com/tidyverse/tidyr/issues/1029), but at this point it is probably slightly better to be backwards compatible.