@DavisVaughan 2021-11-12
@mgirlich would you like to review any further?

At this point I am ready to merge. No revdeps are affected now that wehoop and ffscrapr have already updated their packages on CRAN.