@DavisVaughan 2021-11-12
> the handling of indices as they are not type stable

If we are really looking for some way to justify this, Hadley reminded me that one way to look at this is that an _index_ could be an integer or character vector, so it can be seen as a union type of integer+character. And since `x[<index>]` works correctly with integer or character vectors, it should be fine for most usage. Obviously not a perfect argument, but that + preserving backwards compatibility is enough to convince me to keep this as is.