@mgirlich 2021-11-15
A couple more things then I think it is ready to merge

1. An early return to `col_to_wide()` for data frames would be great. Adding this

```r
  if (is.data.frame(col)) {
    out <- col

    if (!is.null(names_sep)) {
      outer <- name
      inner <- colnames(out)
      names(out) <- apply_names_sep(outer, inner, names_sep)
    }

    return(out)
  }
```

drastically improves the performance

```r
set.seed(1)

n <- 10e3
df <- tibble(
  x = tibble(a = sample(1:n))
)
bench::mark(unnest_wider(df, x))


# Current
# A tibble: 1 × 13
  expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 unnest_wider(df, x)    2.71s    2.71s     0.370     1.5GB     17.0     1    46

# With early return
# A tibble: 1 × 13
  expression               min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 unnest_wider(df, x)   2.04ms   2.27ms      421.    1.93MB     31.4   134    10  
```

2. Similarly an early return to `col_to_long()` for non-list vectors would be great. This passes the tests and is way faster

```r
  if (!vec_is_list(col)) {
    out <- tibble("{name}" := col)

    if (is_false(indices_include)) {
      return(out)
    }

    indices <- vec_names(col)
    all_unnamed <- is.null(indices)

    if (is.null(indices_include) && all_unnamed) {
      return(out)
    }

    if (all_unnamed) {
      out[[indices_to]] <- vec_rep(1L, vec_size(col))
    } else {
      out[[indices_to]] <- indices
    }

    return(out)
  }
```

```r
set.seed(1)

n <- 10e3
df <- tibble(
  x = tibble(a = sample(1:n))
)
bench::mark(unnest_longer(df, x))

# Current
# A tibble: 1 × 13
  expression                min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 unnest_longer(df, x)    212ms    266ms      3.76     3.6MB     9.40     2     5

# With early return
# A tibble: 1 × 13
  expression                min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc
  <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>
1 unnest_longer(df, x)   2.74ms   3.15ms      316.    2.19MB     38.1    91    11
```

3. Can we get rid of the duplicated `New names` notification?
``` r
devtools::load_all("~/GitHub/tidyr/")
#> ℹ Loading tidyr

a <- set_names(1:3, c("x", NA, ""))
# get rid of duplicated new names notification?
tibble(a) %>% 
  unnest_wider(a)
#> New names:
#> * `` -> ...1
#> New names:
#> * `` -> ...1
#> # A tibble: 3 × 2
#>       x  ...1
#>   <int> <int>
#> 1     1    NA
#> 2    NA     2
#> 3    NA     3
```

<sup>Created on 2021-11-15 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>

4. In `unnest_wider()` the `names_sep` doesn't work well with `NA`

``` r
devtools::load_all("~/GitHub/tidyr/")
#> ℹ Loading tidyr

a <- set_names(1:3, c("x", NA, ""))
# "a_NA" name?
tibble(a) %>% 
  unnest_wider(a, names_sep = "_")
#> # A tibble: 3 × 3
#>     a_x  a_NA    a_
#>   <int> <int> <int>
#> 1     1    NA    NA
#> 2    NA     2    NA
#> 3    NA    NA     3
```

<sup>Created on 2021-11-15 by the [reprex package](https://reprex.tidyverse.org) (v2.0.1)</sup>